requirements.txt
loguru
pandas
openpyxl
requests
google-generativeai
PyPDF2
python-jobspy